+++ function.c
@@ -12,33 +12,33 @@ void help(){
 
 
 void testAdd(){
-	int first = 26;
-	int sec = 10;
-	int expected = 36;
-	int actual = add(first, sec);
+	float first = 26;
+	float sec = 10;
+	float expected = 36;
+	float actual = add(first, sec);
 	
 	if(expected != actual){
-		fprintf(stderr,"testAdd has failed! actual: %d, expected: %d", actual, expected);
+		fprintf(stderr,"testAdd has failed! actual: %f, expected: %f", actual, expected);
 		exit(1);
 	}	
 	return;
 }
 void testMultiply(){
- 	int first = 26;
-        int sec = 10;
-        int expected = 260;
-        int actual = multiply(first, sec);
+ 	float first = 26;
+        float sec = 10;
+        float expected = 260;
+        float actual = multiply(first, sec);
 
 
 	if(expected != actual){
-        	fprintf(stderr,"testMultiply has failed! actual: %d, expected: %d", actual, expected);
+        	fprintf(stderr,"testMultiply has failed! actual: %f, expected: %f", actual, expected);
                 exit(1);
         }
         return;
 }
 void testDivide(){
-        int first = 260;
-        int sec = 10;
+        float first = 260;
+        float sec = 10;
         float expected = 26;
         float actual = divide(first, sec);
 
@@ -50,27 +50,27 @@ void testDivide(){
         return;
 }
 void testSubtract(){
-        int first = 26;
-        int sec = 10;
-        int expected = 16;
-        int actual = subtract(first, sec);
+        float first = 26;
+        float sec = 10;
+        float expected = 16;
+        float actual = subtract(first, sec);
 
 
         if(expected != actual){
-                fprintf(stderr,"testSubtract has failed! actual: %d, expected: %d", actual, expected);
+                fprintf(stderr,"testSubtract has failed! actual: %f, expected: %f", actual, expected);
                 exit(1);
         }
         return;
 }
 void testModulus(){
-        int first = 26;
-        int sec = 10;
-        int expected = 6;
-        int actual = modulus(first, sec);
+        float first = 26;
+        float sec = 10;
+        float expected = 6;
+        float actual = modulus(first, sec);
 
 
         if(expected != actual){
-                fprintf(stderr,"testModulus has failed! actual: %d, expected: %d", actual, expected);
+                fprintf(stderr,"testModulus has failed! actual: %f, expected: %f", actual, expected);
                 exit(1);
         }
         return;
